♻️ narrow execution package exports
