+++ libs/backend/agents/execution/admission/main/src/index.ts
@@ -1,1 +1,1 @@
-export { __CreateManagedRunAdmissionPort, __ReadRunAdmissionConcurrencyPolicy } from "./managed-run-admission.js";
+export * from "./managed-run-admission.composition.js";
