+++ libs/backend/agents/execution/runs/main/src/__tests__/attempt-model-key-rejection.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import type { AttemptModelKeyIssuer, AttemptModelKeyMintRequest, MintedAttemptModelKey } from "../run-dispatch.types.js";
+import type { AttemptModelKeyIssuer, AttemptModelKeyMintRequest, MintedAttemptModelKey } from "../attempt-model-key.types.js";
 
 /**
  * Attempt-scoped LiteLLM credential rejection proofs (Phase E slice 4).
